LLM Course/3. Fine-tuning a pretrained model/requirements.txt
torch>=2.0.0
transformers>=4.30.0
datasets>=2.10.0
evaluate>=0.4.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.64.0
pyyaml>=6.0
